requirements.txt
torch
numpy
scipy
pandas
matplotlib
seaborn
scikit-image
scikit-learn
tqdm
pyyaml
imageio
ezdxf
